feat: separate toolbar and navbar
